test/fixtures/wpt/FileAPI/url/unicode-origin.sub.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>FileAPI Test: Verify origin of Blob URL</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<body>
<script>
async_test(t => {
  const frame = document.createElement('iframe');
  self.addEventListener('message', t.step_func(e => {
      if (e.source != frame.contentWindow) return;
      const url = e.data.url;
      assert_false(url.includes('天気の良い日'),
          'Origin should be ascii rather than unicode');
      assert_equals(new URL(url).origin, e.origin,
          'Origin of URL should match origin of frame');
      assert_true(url.startsWith('blob:{{location[scheme]}}://xn--'));
      t.done();
    }));
  frame.src = '{{location[scheme]}}://{{domains[天気の良い日]}}:{{location[port]}}/FileAPI/support/url-origin.html';
  document.body.appendChild(frame);
}, 'Verify serialization of non-ascii origin in Blob URLs');
</script>
